all new :)
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Retreat2008@4219 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/CONCEPT/Exception.h
@@ -248,7 +248,7 @@ namespace OpenMS
 			: public BaseException
 		{
 			public:
-				IndexOverflow(const char* file, int line, const char* function, Int index = 0, UInt size = 0) throw();
+				IndexOverflow(const char* file, int line, const char* function, Int index = 0, size_t size = 0) throw();
 		};
 
 		/**
